+++ ChangeLog.md
@@ -31,6 +31,7 @@ This release contains multiple breaking changes. Please read the [upgrade guide]
 - (Import) Support for up to 8 UV sets (note: glTF shaders still support only two sets; part of #206)
 - `IMaterialGenerator` was extended with support for points topology
 - (Export) `GameObjectExportSettings.disabledComponents` to explicitely enable export of disabled components (e.g. `MeshRenderer`, `Camera`, or `Light`)
+- (Export) `ExportSettings.componentMask` to include or exclude components from export based on type
 ### Changed
 - Converted a lot of unintentionally public classes, types and properties to internal ones
 - Replaced `CollectingLogger.item` with `.Count` and `.Items` iterator